feat: move fixtures to conftest
